Moved showSettings
